feat: rearrange testing so tests and plugin work
