+++ pyproject.toml
@@ -36,4 +36,4 @@ select = [
 convention = "google"
 
 [tool.ruff.per-file-ignores]
-"test/*.py" = ["D103"]
+"**/test/*.py" = ["D103"]
